Removed memcached dependency from the charm for coordination.
There is still discussion going on in charms community not to use the
memcached as a backend coordinator because it is not recommended by
OpenStack and the designate coordination code requires heartbeats,
locks, etc. which is not supported by memcached. So there will probably
be more changes coming in the charm to either use redis, etcd or zookeeper
for coordination backend.

Refrence: https://bugs.launchpad.net/charm-designate/+bug/1759597